+++ apps/pt_tvmdsoop/prepare_and_test_pt_tvm_class.sh
@@ -34,10 +34,11 @@ fi
 python3 -c "import tvm; print(tvm.runtime.enabled('gpu'))" | grep -e 1
 
 if [ "$?" -eq 0 ]; then
-    echo "Build PT_TVMCLASS with gpu support and execute tests"
+    echo "Build PT_TVMDSOOP with gpu support and execute tests"
     CMAKE_OPTIONS="-DUSE_CUDA=ON -DUSE_CUDNN=ON -DPython3_EXECUTABLE=python3 -DTVM_ROOT=${TVM_ROOT}"
     mkdir -p build
     cd build; cmake .. ${CMAKE_OPTIONS} && make
+    cp *.so $TVM_ROOT/build/
     cd ..
 
     LD_LIBRARY_PATH=${TVM_ROOT}/build:./build:$LD_LIBRARY_PATH python3 -m pytest -v ./tests
