+++ cmake/modules/LibInfo.cmake
@@ -60,7 +60,7 @@ function(add_lib_info src_file)
     TVM_INFO_INSTALL_DEV="${INSTALL_DEV}"
     TVM_INFO_HIDE_PRIVATE_SYMBOLS="${HIDE_PRIVATE_SYMBOLS}"
     TVM_INFO_USE_TF_TVMDSOOP="${USE_TF_TVMDSOOP}"
-    TVM_INFO_USE_PT_TVMCLASS="${USE_PT_TVMCLASS}"
+    TVM_INFO_USE_PT_TVMDSOOP="${USE_PT_TVMDSOOP}"
     TVM_INFO_USE_FALLBACK_STL_MAP="${USE_FALLBACK_STL_MAP}"
     TVM_INFO_USE_BYODT_POSIT="${USE_BYODT_POSIT}"
     TVM_INFO_USE_BLAS="${USE_BLAS}"
